% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.r
\name{cv}
\alias{cv}
\title{Coefficient of Variance}
\usage{
cv(object)
}
\arguments{
\item{object}{is a daily or monthly precipitation serie.}
}
\value{
the coefficient of variance
}
\description{
It calculates the coefficient of variance of a monthly precipitation.
}
\details{
A daily precipitation serie is transformed to a monthly serie.
}
\examples{
##
# Loading the montly precipitation serie.
data(monthly)

##
# Calculating the Coefficient of Variance
cv(monthly)
}
\author{
Lucas Venezian Povoa \email{lucasvenez@gmail.com}
}

