% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/precintcon.r
\docType{data}
\name{daily}
\alias{daily}
\title{Daily precipitation between 1976 and 2010}
\format{A data frame with 420 observations on the following 33 variables:
\describe{
   \item{\code{year}}{the year.}
   \item{\code{month}}{the month.}
   \item{\code{d1}}{the precipitation value in millimeters of the day #1 of the month.}
   \item{\code{d2}}{the precipitation value in millimeters of the day #2 of the month.}
   \item{\code{d3}}{the precipitation value in millimeters of the day #3 of the month.}
   \item{\code{d4}}{the precipitation value in millimeters of the day #4 of the month.}
   \item{\code{d5}}{the precipitation value in millimeters of the day #5 of the month.}
   \item{\code{d6}}{the precipitation value in millimeters of the day #6 of the month.}
   \item{\code{d7}}{the precipitation value in millimeters of the day #7 of the month.}
   \item{\code{d8}}{the precipitation value in millimeters of the day #8 of the month.}
   \item{\code{d9}}{the precipitation value in millimeters of the day #9 of the month.}
   \item{\code{d10}}{the precipitation value in millimeters of the day #10 of the month.}
   \item{\code{d11}}{the precipitation value in millimeters of the day #11 of the month.}
   \item{\code{d12}}{the precipitation value in millimeters of the day #12 of the month.}
   \item{\code{d13}}{the precipitation value in millimeters of the day #13 of the month.}
   \item{\code{d14}}{the precipitation value in millimeters of the day #14 of the month.}
   \item{\code{d15}}{the precipitation value in millimeters of the day #15 of the month.}
   \item{\code{d16}}{the precipitation value in millimeters of the day #16 of the month.}
   \item{\code{d17}}{the precipitation value in millimeters of the day #17 of the month.}
   \item{\code{d18}}{the precipitation value in millimeters of the day #18 of the month.}
   \item{\code{d19}}{the precipitation value in millimeters of the day #19 of the month.}
   \item{\code{d20}}{the precipitation value in millimeters of the day #20 of the month.}
   \item{\code{d21}}{the precipitation value in millimeters of the day #21 of the month.}
   \item{\code{d22}}{the precipitation value in millimeters of the day #22 of the month.}
   \item{\code{d23}}{the precipitation value in millimeters of the day #23 of the month.}
   \item{\code{d24}}{the precipitation value in millimeters of the day #24 of the month.}
   \item{\code{d25}}{the precipitation value in millimeters of the day #25 of the month.}
   \item{\code{d26}}{the precipitation value in millimeters of the day #26 of the month.}
   \item{\code{d27}}{the precipitation value in millimeters of the day #27 of the month.}
   \item{\code{d28}}{the precipitation value in millimeters of the day #28 of the month.}
   \item{\code{d29}}{the precipitation value in millimeters of the day #29 of the month.}
   \item{\code{d30}}{the precipitation value in millimeters of the day #30 of the month.}
   \item{\code{d31}}{the precipitation value in millimeters of the day #31 of the month.}
 }}
\source{
National Water Agency (ANA) \url{http://www2.ana.gov.br/Paginas/default.aspx}
}
\usage{
data(daily)
}
\description{
This data set gives the daily precipitation (in millimeters) between 1976 and 2010, as compiled by the National Water Agency (ANA), Brazil.
}
\examples{
data(daily)
## maybe pci(daily) ; rai(daily) ...
}
\keyword{datasets}

