% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ff.index.r
\name{ff.index}
\alias{ff.index}
\title{F factor}
\usage{
ff.index(object)
}
\arguments{
\item{object}{is a daily or monthly precipitation serie}
}
\value{
the ff index in millimeters
}
\description{
It calculates the Approximated Rainfall Erosivity Factor according to the ff index.
}
\examples{
##
# Loading the daily precipitation serie.
data(daily)

##
# Calculating ff index
ff.index(daily)
}
\author{
Lucas Venezian Povoa
}
\references{
Ferro, V., Giordano, G., Iovino, M. (1991). Isoerosivity and Erosion Risk Map for Sicily.
Hydrolog. Sci. J. 36, 549-564
}

