% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pci.r
\name{pci}
\alias{pci}
\alias{precintcon.pci.analysis}
\title{Precipitation Concentration Index}
\usage{
pci(object)
}
\arguments{
\item{object}{a daily or monthly precipitation serie.}
}
\value{
A data.frame containing the following variables:
\itemize{
\item \code{year} is the year.
\item \code{pci} is the precipitation concentration index.
}
}
\description{
It calculates the Precipitation Concentration Index (PCI) on a 
daily or monthly precipitation serie.
}
\examples{
##
# Loading the monthly precipitation serie.
data(monthly)

## 
# Performing the Precipitation Concentration Index analysis
pci(monthly)
}
\author{
Lucas Venezian Povoa \email{lucasvenez@gmail.com}
}
\seealso{
\code{\link{pplot.pci}}
\code{\link{read.data}}
\code{\link{as.daily}}
\code{\link{as.monthly}}
}
\keyword{concentration}
\keyword{index}
\keyword{precipitation}

