% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/precintcon.r
\docType{package}
\name{precintcon}
\alias{precintcon}
\alias{precintcon-package}
\title{Precipitation Intensity, Concentration and Anomaly Analysis}
\description{
It contains functions to analyze the precipitation
intensity, concentration and anomaly.
}
\author{
Lucas Venezian Povoa \email{lucasvenez@gmail.com}
}
\references{
Michael J. Hayes (2015). \emph{Comparison of Major Drought Indices}.
National Drought Mitigation Center. \url{http://goo.gl/skHfs9}.

Gibbs, W. J. (1967). \emph{Rainfall deciles as drought indicators}.
Bureau of Meteorology Bulletin, n. 48. Commonwealth of Australia.

Luis, M. D., Gonzolez-Hidalgo, J. C., Brunetti, M., Longares, L. A. (2011).
\emph{Precipitation concentration changes in Spain 1946-2005}.
Natural Hazards and Earth System Science, v. 11, n. 5, p. 1259--1265.

Martin-Vide, J. (2004). \emph{Spatial distribution of a daily precipitation
concentration index in peninsular Spain}. International Journal of Climatology,
v. 24, n. 8, p. 959--971.

Guttman, N. B. (1999). \emph{Accepting the standardized precipitation index: a
calculation algorithm}. Journal of the American Water Resources Association, v. 35,
n. 2, p. 311--322. Blackwell Publishing Ltd.

Barring, L., Hulme, M. (1991). \emph{Filters and approximate confidence intervals for
interpreting rainfall anomaly indices}. Journal of climate, v. 4, n. 8, p. 837--847.
}
\keyword{anomaly}
\keyword{concentration}
\keyword{intensity}
\keyword{package}
\keyword{precipitation}
\keyword{rainfall}

