% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tii.r
\name{tii}
\alias{tii}
\title{Temporaly Irregularity Index}
\usage{
tii(object)
}
\arguments{
\item{object}{is a daily or monthly precipitation serie}
}
\value{
the temporaly irregularity index according to the function sum(log(Pi+1/Pi))/(n-1)
}
\description{
It calculates the temporaly irregularity index according to the function sum(log(Pi+1/Pi))/(n-1),
where Pi is the precipitation amount of year i, and n is the number of years.
}
\details{
Daily or monthly precipitation series are transformed to annual series.
}
\examples{
##
# Loading the monthly precipitation serie
data(monthly)

##
# Calculationg the Temporaly Irregularity Index
tii(monthly)
}
\author{
Lucas Venezian Povoa
}

