% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.test.r
\name{trend.test}
\alias{trend.test}
\title{Mann-Kendall Trend Test}
\usage{
trend.test(object, significance.level = 0.05)
}
\arguments{
\item{object}{is a daily or monthly precipitation serie.}

\item{significance.level}{is the significance level to be considered in the analysis. 
It is usually equals to 5\% (default: 0.05).}
}
\value{
A trend test index.
}
\description{
It calculates a trend test using the rank-based nonparametric Mann-Kendall method.
}
\author{
Lucas Venezian Povoa \email{lucasvenez@gmail.com}
}
\references{
H. B. Mann (1945). Nonparametric tests against trend. Econometrica, vol. 13, pp. 245-259.
M. G. Kendall (1975). Rank Correlation Methods. Griffin, London, UK.
}
\seealso{
\code{\link{ci}} \code{\link{ci.per.year}} \code{\link{pci}} \code{\link{read.data}}
}
\keyword{precipitation}
\keyword{test}
\keyword{trend}

